src/requirements.txt
discord.py==2.3.2
# openai==1.3.0
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
zeep==4.2.1
